Refactored code

+++ src/main/java/hello/domains/User.java
@@ -1,6 +1,5 @@
 package hello.domains;
 
-import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -12,12 +11,12 @@ public class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     private String name;
     private Integer age;
     private Integer salary;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
